Added another GUI
The second GUI within "buildingGUI.py" will be the expanded GUI with more features and a live camera view within. Adjustments need to be made to the simplistic one to ensure that when the button is clicked, the camera can open and the computer vision can take place within both documents.